Move compatibility checks to SignificantBuilds
